requirements.txt
streamlit
langchain-core
langchain-openai
langchain-community
supabase
unstructured[pdf]
cohere
pydantic
python-dotenv